// types.ts
export enum UserRole {
  ADMIN = 'admin',
  EMPLOYEE = 'employee',
}

export interface User {
  id: string;
  username: string;
  email: string;
  role: UserRole;
  firstName: string; 
  lastName: string;  
  profilePictureUrl?: string | null; 
  phone?: string;
  department?: string;
  joinDate?: string;
  reportsTo?: string | null;
}
// ...
export type AdminUserUpdateData = Pick<User, 'firstName' | 'lastName' | 'email' | 'phone' | 'department' | 'joinDate' | 'role' | 'profilePictureUrl' | 'reportsTo'>;